Move parallel_exec() to exectools; keep only one version
